requirements.txt
python-bitcoinlib
pyzmq
BeautifulSoup4